GH-17: moving settings store
